etl/requirements.txt
pandas==2.2.3
requests==2.32.3
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
